pyproject.toml
[project]
name = "langchain-learn"
version = "0.1.0"
description = ""
authors = [
    {name = "Ojie Precious",email = "ojiepreshuz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"

[tool.poetry]
package-mode = false 

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
langchain = "^1.0.0"
langchain-community = "^0.4.0"
langchain-openai = "^1.0.0"
langchain-anthropic = "^0.1.15"
langchain-google-genai = "^3.0.0"
langchain-google-firestore = "^0.3.0"
python-dotenv = "^1.0.1"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
sentence-transformers = "^3.0.0"
bs4 = "^0.0.2"
firecrawl-py = "^0.0.13"
langchainhub = "^0.1.18"
wikipedia = "^1.4.0"
tavily-python = "^0.3.3"
firestore = "^0.0.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
